# apply: improve performance of `moving`